A first attempt at working with GENCODE and RefSeq
Using bed files for GENCODE and RefSeq as well as a file for each, linking the gene symbols to respective gene IDs. However, I noticed that the gene Ids for GENCODE are actually transcript IDs. That combined with too many input files for the databases, lead to this version being discarded. But maybe its needed in the future for parts, thus the commit.